Split out the tests for the two managers, they are getting pretty large.
