+++ managers/manager_memory.go
@@ -386,7 +386,6 @@ func (cm ContentManagerMemory) ListAllTags(page int, perPage int) (*models.Tags,
     return nil, errors.New("Not implemented")
 }
 
-// hate
 func (cm ContentManagerMemory) ListAllTagsContext() (*models.Tags, error) {
     _, limit, page := GetPagination(cm.Params(), cm.cfg.Limit)
     return cm.ListAllTags(page, limit)
